Move B2CTrustedHostList and setKnownAuthorities
