+++ lib/msal-core/src/utils/Constants.ts
@@ -107,8 +107,6 @@ export const AADTrustedHostList =  {
     "login.microsoftonline.us": "login.microsoftonline.us"
 };
 
-export const B2CTrustedHostList = {};
-
 /**
  * @hidden
  * SSO Types - generated to populate hints
